backend/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
google-generativeai